Spinning out `entity::Player` module and moving init/reset code there
